Define compose-destinations
