+++ app/src/main/java/pl/mateuszteteruk/starwars/StarWarsApp.kt
@@ -1,0 +1,17 @@
+package pl.mateuszteteruk.starwars
+
+import androidx.compose.runtime.Composable
+import com.ramcosta.composedestinations.DestinationsNavHost
+import pl.mateuszteteruk.starwars.movie.NavGraphs
+import pl.mateuszteteruk.starwars.ui.theme.StarWarsTheme
+
+@Composable
+fun StarWarsApp() {
+    StarWarsTheme {
+        DestinationsNavHost(
+            navGraph = NavGraphs.root,
+            startRoute = NavGraphs.root.startRoute,
+        )
+    }
+}
+
